Move skills section into components folder
